CommentProviderとCommentDataを分離
