+++ Assets/TechC/VBattle/Scripts/InGame/Comment/BaffSystem/CommentData.cs
@@ -1,0 +1,26 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace TechC
+{
+   /// <summary>
+    /// コメントタイプを定義
+    /// </summary>
+    public enum CommentType { Normal, Buff, MapChange }
+
+    public class CommentData
+    {
+        public CommentType type;
+        public string text;
+        public BuffType? buffType;
+
+        /* コンストラクタでコメントタイプ、テキスト、バフタイプを設定 */
+        public CommentData(CommentType type, string text, BuffType? buffType)
+        {
+            this.type = type;
+            this.text = text;
+            this.buffType = buffType;
+        }
+    }
+}
