+++ Assets/TechC/VBattle/Scripts/InGame/Comment/BaffSystem/CommentData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: af632686a80523b489a62b9a7b4abf30
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
