+++ Assets/TechC/VBattle/Scripts/InGame/Comment/CommentSystem/CommentProvider.cs
@@ -4,25 +4,6 @@ using UnityEngine;
 namespace TechC
 {
     /// <summary>
-    /// コメントタイプを定義
-    /// </summary>
-    public enum CommentType { Normal, Buff, MapChange }
-
-    public class CommentData
-    {
-        public CommentType type;
-        public string text;
-        public BuffType? buffType;
-
-        /* コンストラクタでコメントタイプ、テキスト、バフタイプを設定 */
-        public CommentData(CommentType type, string text, BuffType? buffType)
-        {
-            this.type = type;
-            this.text = text;
-            this.buffType = buffType;
-        }
-    }
-    /// <summary>
     /// ランダムにコメントを選び提供する
     /// </summary>
     public class CommentProvider : MonoBehaviour
